Moved more functions to pipeline.projects.provenance.util.
